Separate and cleaning code
