+++ doc/Makefile.am
@@ -8,7 +8,7 @@ all-local: doc
 # Public programs ("main" tags in project.xml), auto-regenerated:
 MAN1 = fty-nut.1 fty-nut-command.1 fty-nut-configuration.1 fty-nut-configurator.1
 # Public classes ("class" tags in project.xml), auto-regenerated:
-MAN3 = fty_nut_server.3 fty_nut_command_server.3 fty_nut_configurator_server.3 fty_nut_configuration_server.3 alert_actor.3 sensor_actor.3
+MAN3 = fty_nut_server.3 fty_nut_command_server.3 fty_nut_configurator_server.3 fty_nut_configuration_helper.3 fty_nut_configuration_manager.3 fty_nut_configuration_connector.3 fty_nut_configuration_server.3 alert_actor.3 sensor_actor.3
 # Project overview, written by a human after initial skeleton:
 # NOTE: stub doc/fty-nut.adoc is generated by GSL from project.xml
 #       and then comitted to SCM and maintained manually to describe the
@@ -75,6 +75,18 @@ fty_nut_command_server.txt: $(top_srcdir)/src/fty_nut_command_server.cc
 GENERATED_DOCS += fty_nut_configurator_server.txt fty_nut_configurator_server.doc
 fty_nut_configurator_server.txt: $(top_srcdir)/src/fty_nut_configurator_server.cc
 	"$(srcdir)/mkman" "fty_nut_configurator_server" "$(builddir)/fty_nut_configurator_server.txt" "$(srcdir)/.."
+
+GENERATED_DOCS += fty_nut_configuration_helper.txt fty_nut_configuration_helper.doc
+fty_nut_configuration_helper.txt: $(top_srcdir)/src/fty_nut_configuration_helper.cc
+	"$(srcdir)/mkman" "fty_nut_configuration_helper" "$(builddir)/fty_nut_configuration_helper.txt" "$(srcdir)/.."
+
+GENERATED_DOCS += fty_nut_configuration_manager.txt fty_nut_configuration_manager.doc
+fty_nut_configuration_manager.txt: $(top_srcdir)/src/fty_nut_configuration_manager.cc
+	"$(srcdir)/mkman" "fty_nut_configuration_manager" "$(builddir)/fty_nut_configuration_manager.txt" "$(srcdir)/.."
+
+GENERATED_DOCS += fty_nut_configuration_connector.txt fty_nut_configuration_connector.doc
+fty_nut_configuration_connector.txt: $(top_srcdir)/src/fty_nut_configuration_connector.cc
+	"$(srcdir)/mkman" "fty_nut_configuration_connector" "$(builddir)/fty_nut_configuration_connector.txt" "$(srcdir)/.."
 
 GENERATED_DOCS += fty_nut_configuration_server.txt fty_nut_configuration_server.doc
 fty_nut_configuration_server.txt: $(top_srcdir)/src/fty_nut_configuration_server.cc
