+++ include/fty_nut_configuration_server.h
@@ -28,47 +28,64 @@ namespace fty
 {
 namespace nut
 {
+    
+using namespace DBAssetsDiscovery;
+    
+typedef struct ComputeAssetConfigurationUpdateResult {
+    /// \brief Known, working configuration.
+    nutcommon::DeviceConfigurations workingConfigurations;
+    /// \brief Known, non-working configuration.
+    nutcommon::DeviceConfigurations nonWorkingConfigurations;
+    /// \brief Unknown, working configuration.
+    nutcommon::DeviceConfigurations newConfigurations;
+    /// \brief Unknown, unassessable configuration.
+    nutcommon::DeviceConfigurations unknownStateConfigurations;
+} ComputeAssetConfigurationUpdateResult;
 
-class ConfigurationManager
-{
-    public:
-        ConfigurationManager(const std::string& dbConn);
-        ~ConfigurationManager() = default;
+/**
+ * \brief Scan asset for NUT driver configurations.
+ *
+ * The scan will detect the following drivers:
+ * - netxml-ups
+ * - snmp-ups (SNMPv1 and SNMPv3)
+ * - snmp-ups-dmf (SNMPv1 and SNMPv3)
+ *
+ * \warning This won't return the list of all *working* NUT device configurations, as the list of handled NUT drivers is not exhaustive!
+ *
+ * \param pool PoolWorker to use.
+ * \param asset fty_proto_t of asset to scan.
+ * \param credentialsSnmpV1 SNMPv1 credentials to test.
+ * \param credentialsSnmpV3 SNMPv3 credentials to test.
+ * \return All detected and working NUT device configurations.
+ */
+nutcommon::DeviceConfigurations assetScanDrivers(messagebus::PoolWorker& pool, fty_proto_t *asset, const std::vector<nutcommon::CredentialsSNMPv1>& credentialsSnmpV1, const std::vector<nutcommon::CredentialsSNMPv3>& credentialsSnmpV3);
 
-        void automaticAssetConfigurationPrioritySort(fty_proto_t* asset);
-        void scanAssetConfigurations(fty_proto_t* asset);
-        void applyAssetConfiguration(fty_proto_t* asset);
+/**
+ * \brief Sort NUT driver configurations into categories from known and detected configurations.
+ * \param knownConfigurations Known NUT device configurations in database.
+ * \param detectedConfigurations Detected NUT device configurations at runtime.
+ * \return All NUT driver configurations sorted into categories.
+ */
+ComputeAssetConfigurationUpdateResult computeAssetConfigurationUpdate(const nutcommon::DeviceConfigurations& knownConfigurations, const nutcommon::DeviceConfigurations& detectedConfigurations);
 
-    private:
-        messagebus::PoolWorker m_poolScanners;
-        std::string m_dbConn;
-};
+nutcommon::DeviceConfigurations instanciateDatabaseConfigurations(const DeviceConfigurationInfos& dbConfs, fty_proto_t* asset, const std::vector<nutcommon::CredentialsSNMPv1>& credentialsSNMPv1, const std::vector<nutcommon::CredentialsSNMPv3>& credentialsSNMPv3);
 
-class ConfigurationConnector
-{
-    public:
-        struct Parameters {
-            Parameters();
+/**
+ * \brief Find which device configuration type a given device configuration best matches.
+ * \param asset Asset to check with.
+ * \param configuration Device configuration to match.
+ * \param types Device configuration types to match against.
+ * \return Iterator to best device configuration type match, or end of collection if no suitable match found.
+ */
+DeviceConfigurationInfoDetails::const_iterator matchDeviceConfigurationToBestDeviceConfigurationType(fty_proto_t* asset, const nutcommon::DeviceConfiguration& configuration, const DeviceConfigurationInfoDetails& types);
 
-            std::string endpoint;
-            std::string agentName;
-
-            std::string dbUrl;
-        };
-
-        ConfigurationConnector(Parameters params);
-        ~ConfigurationConnector() = default;
-
-    private:
-        void handleRequest(messagebus::Message msg);
-        void sendReply(const messagebus::MetaData& metadataRequest, bool status, const messagebus::UserData& dataReply);
-
-        Parameters m_parameters;
-        ConfigurationManager m_manager;
-        messagebus::Dispatcher<std::string, std::function<messagebus::UserData(messagebus::UserData)>, std::function<messagebus::UserData(const std::string&, messagebus::UserData)>> m_dispatcher;
-        messagebus::PoolWorker m_worker;
-        std::unique_ptr<messagebus::MessageBus> m_msgBus;
-};
+/**
+ * \brief Return the order of preference for an asset's driver configurations.
+ * \param asset Asset to sort device configurations with.
+ * \param configurations Device configurations to sort.
+ * \return List of indexes of driver configurations, ordered from most to least preferred.
+ */
+std::vector<size_t> sortDeviceConfigurationPreferred(fty_proto_t* asset, const nutcommon::DeviceConfigurations& configurations);
 
 }
 }
