+++ include/fty_nut_library.h
@@ -111,6 +111,12 @@ typedef struct _fty_nut_command_server_t fty_nut_command_server_t;
 #define FTY_NUT_COMMAND_SERVER_T_DEFINED
 typedef struct _fty_nut_configurator_server_t fty_nut_configurator_server_t;
 #define FTY_NUT_CONFIGURATOR_SERVER_T_DEFINED
+typedef struct _fty_nut_configuration_helper_t fty_nut_configuration_helper_t;
+#define FTY_NUT_CONFIGURATION_HELPER_T_DEFINED
+typedef struct _fty_nut_configuration_manager_t fty_nut_configuration_manager_t;
+#define FTY_NUT_CONFIGURATION_MANAGER_T_DEFINED
+typedef struct _fty_nut_configuration_connector_t fty_nut_configuration_connector_t;
+#define FTY_NUT_CONFIGURATION_CONNECTOR_T_DEFINED
 typedef struct _fty_nut_configuration_server_t fty_nut_configuration_server_t;
 #define FTY_NUT_CONFIGURATION_SERVER_T_DEFINED
 typedef struct _alert_actor_t alert_actor_t;
@@ -123,6 +129,9 @@ typedef struct _sensor_actor_t sensor_actor_t;
 #include "fty_nut_server.h"
 #include "fty_nut_command_server.h"
 #include "fty_nut_configurator_server.h"
+#include "fty_nut_configuration_helper.h"
+#include "fty_nut_configuration_manager.h"
+#include "fty_nut_configuration_connector.h"
 #include "fty_nut_configuration_server.h"
 #include "alert_actor.h"
 #include "sensor_actor.h"
